Refactored pipeline and implemented MuSE (resolves #32)
resolves #32

Pipeline completely refactored into a better library-like structure.
Mutation calls happen on per-chromosome basis.
