+++ src/protect/mutation_calling/indel.py
@@ -12,5 +12,18 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from __future__ import print_function
 
-from __future__ import absolute_import
+
+def run_indel_caller(job, tumor_bam, normal_bam, univ_options, indel_options):
+    """
+    This module will run an indel caller on the DNA bams.  This module will be
+    implemented in the future.
+
+    This module corresponds to node 13 on the tree
+    """
+    job.fileStore.logToMaster('Running INDEL on %s' % univ_options['patient'])
+    job.fileStore.logToMaster('INDELs are currently unsupported.... Skipping.')
+    indel_file = job.fileStore.getLocalTempFile()
+    output_file = job.fileStore.writeGlobalFile(indel_file)
+    return output_file
